server/requirements.txt
asgiref==3.4.1
Django==3.2.7
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
numpy==1.21.2
opencv-python==4.5.3.56
Pillow==8.3.2
PyJWT==1.7.1
pytz==2021.1
sqlparse==0.4.2
